fix: port decryptFromkeystore
